Add an initial CODEOWNERS file to the pygame-ce project
I hope this will allow us to have more control over reviewing pull requests and let us expand the pool of reviewers in a steady and encouraging way. The idea being that it can become relatively easy to become an 'impactful reviewer' where an approving review effects whether a PR can be merged or not. While at the same time we don't expose too much risk of allowing free-reign for a couple of inexperienced reviewers accidentally making a mess of the main branch.