Allow vsync without OpenGL or SCALED flag
This is a work in progress. Right now we force a SDL_Renderer/GPU code
path when vsync without other flags indicating GPU is passed.

Resizing and fullscreen work, but toggling fullscreen does not. The
behaviour is meant to be indistinguishable from regular un-SCALED pygame
software rendering. Re-sizing the window changes the size of the display
surface. This feels like a horrible hack, but it runs smooth