fixed a bug
For a split second the page would look like light theme and then it would load the js file and quickly change the class attribute of the body tag. This was an annoying thing and I fixed it by using the <html> tag for putting the class and removing the `defer` attribute when loading the script.