Add dark theme for documentation (#2122)
* Add dark theme for documentation

* Remove custom scrollbar

* remove console.log()

* Correct exception handling and paths in makeref.py

Didn't realize what I was doing was dumb

I copied the code from @gresm  from https://github.com/pygame-community/pygame-ce/pull/2122#issuecomment-1540584050

* moved theme icons to their own files, space between assignment

* proper way of adding js files

Thanks to @gresm for telling me the correct way to add it.

* get rid of unused library & add blank line

* revert to single quotes

* revert to single quotes

* add newline

* move searchbar to svg files, fix searchbar button not working, reverse the way data-theme works

* scale down icons

* reverse theme change icons

* get rid of useless theme search icon

* Update pygame.css_t

* fixed a bug

For a split second the page would look like light theme and then it would load the js file and quickly change the class attribute of the body tag. This was an annoying thing and I fixed it by using the <html> tag for putting the class and removing the `defer` attribute when loading the script.