Use HasColorKey instead of GetColorKey in pgSurface_Blit
This routine doesn't use the value of GetColorKey, it just wants to know whether the surface has one.

Internally, this was setting a bunch of SDL errors ("surf doesn't have colorkey"-esque), which became much more expensive in SDL 2.29.3 because of a new logging feature.

Therefore lets replace it with HasColorKey (a newer function, the code was originally written before this existed). This fixes the performance regression here and even makes the performance slightly better than it was before, at least in my test scenario.

Credit to itzpr for narrowing this down to SDL_GetColorKey