Remove unnecessary pixeloffset member of subsurface data
I was digging around our structs and found this member barely used anywhere. All it did was set the subsurf pixel pointer to the correct location when pgSurface_Prep is called. But the subsurf pixel pointer is set up to point at the right location in surface.subsurface already, and never changes from that position. So storing the offset and re-setting it occasionally is unnecessary effort.