Adjust pg_(Get/Set)DefaultConvertFormat for SDL3
This reworks the functions to work on pixel format enums instead of structs, because in SDL3 the enums are much easier to get than the structs.

This required involved changes to surface.c, but I think the previous mask logic is much clearer now that it is expressed using format enums.