Fix Rect contains error message
This error shouldn't exist at all imo, but for backwards compatibility lets keep it for now. And the error should at least be accurate, so this commit fixes that.

Previously it would say "int" even on FRects, where it meant float. Also strengthens the tests to explicitly test the TypeError behavior of this mismatched number type scenario.